# openapi.json
{"openapi":"3.0.3","info":{"title":"Books API","description":"A simple Books management API","version":"1.0.0"},"paths":{"/books":{"get":{"tags":["books"],"summary":"Read All Books","operationId":"read_all_books_books_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/books/{book_id}":{"get":{"tags":["books"],"summary":"Read Book","operationId":"read_book_books__book_id__get","parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"description":"The ID of the book to retrieve","title":"Book Id"},"description":"The ID of the book to retrieve"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["books"],"summary":"Delete Book","operationId":"delete_book_books__book_id__delete","parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"title":"Book Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/":{"get":{"tags":["books"],"summary":"Read Book By Rating","operationId":"read_book_by_rating_books__get","parameters":[{"name":"book_rating","in":"query","required":true,"schema":{"type":"integer","exclusiveMaximum":6,"exclusiveMinimum":0,"description":"Filter by rating (1-5)","title":"Book Rating"},"description":"Filter by rating (1-5)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/publish/":{"get":{"tags":["books"],"summary":"Read Books By Publish Date","operationId":"read_books_by_publish_date_books_publish__get","parameters":[{"name":"published_date","in":"query","required":true,"schema":{"type":"integer","exclusiveMaximum":2031,"exclusiveMinimum":1999,"title":"Published Date"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/create-book":{"post":{"tags":["books"],"summary":"Create Book","operationId":"create_book_create_book_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/update_book":{"put":{"tags":["books"],"summary":"Update Book","operationId":"update_book_books_update_book_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookRequest"}}},"required":true},"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BookRequest":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id","description":"ID is not needed on create"},"title":{"type":"string","minLength":3,"title":"Title"},"author":{"type":"string","minLength":1,"title":"Author"},"description":{"type":"string","maxLength":100,"minLength":1,"title":"Description"},"rating":{"type":"integer","exclusiveMaximum":6.0,"exclusiveMinimum":0.0,"title":"Rating"},"published_date":{"type":"integer","exclusiveMaximum":2031.0,"exclusiveMinimum":1999.0,"title":"Published Date"}},"type":"object","required":["title","author","description","rating","published_date"],"title":"BookRequest","example":{"author":"codingwithroby","description":"A new description of a book","published_date":2029,"rating":5,"title":"A new book"}},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}